refactor(validator): broke up layout.go into layout & zones
